Refactoring the image editor
