+++ concrete/elements/files/edit/image.php
@@ -2,13 +2,13 @@
 
 defined('C5_EXECUTE') or die('Access Denied.');
 
-use Concrete\Core\ImageEditor\EditorServiceProvider;
+use Concrete\Core\ImageEditor\ImageEditorService;
 
 /**
  * @var Concrete\Core\Entity\File\Version $fv
  * @var Concrete\Core\Application\Application $app
  */
 
-/** @var EditorServiceProvider $editorService */
-$editorService = $app->make(EditorServiceProvider::class);
+/** @var ImageEditorService $editorService */
+$editorService = $app->make(ImageEditorService::class);
 $editorService->renderActiveEditor($fv);
